move myApp to another file and update codecov
