+++ lib/initial_app.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:hatspace/features/home/view/home_view.dart';
+import 'package:hatspace/models/authentication/authentication_bloc.dart';
+import 'package:hatspace/strings/l10n.dart';
+import 'package:hatspace/theme/hs_theme.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: lightThemeData,
+      home: MultiBlocProvider(
+        providers: [
+          BlocProvider<AuthenticationBloc>(
+            create: (context) => AuthenticationBloc(),
+          )
+          // TODO add your bloc creation here
+        ],
+        child: const HomePageView(),
+      ),
+      localizationsDelegates: const <LocalizationsDelegate<dynamic>>[
+        HatSpaceStrings.delegate
+      ],
+      supportedLocales: HatSpaceStrings.delegate.supportedLocales,
+    );
+  }
+}
