+++ lib/main.dart
@@ -1,13 +1,9 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:hatspace/features/home/view/home_view.dart';
 import 'package:hatspace/firebase_options.dart';
-import 'package:hatspace/models/authentication/authentication_bloc.dart';
+import 'package:hatspace/main_app.dart';
 import 'package:hatspace/singleton/hs_singleton.dart';
-import 'package:hatspace/strings/l10n.dart';
-import 'package:hatspace/theme/hs_theme.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -22,29 +18,4 @@ void main() async {
       options: DefaultFirebaseOptions.currentPlatform);
 
   runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: lightThemeData,
-      home: MultiBlocProvider(
-        providers: [
-          BlocProvider<AuthenticationBloc>(
-            create: (context) => AuthenticationBloc(),
-          )
-          // TODO add your bloc creation here
-        ],
-        child: const HomePageView(),
-      ),
-      localizationsDelegates: const <LocalizationsDelegate<dynamic>>[
-        HatSpaceStrings.delegate
-      ],
-      supportedLocales: HatSpaceStrings.delegate.supportedLocales,
-    );
-  }
 }
